requirements.txt
PyYAML>=6.0
jsonschema>=4.20
requests>=2.28
# Google API (Sheets + BigQuery)
gspread>=5.0
google-auth>=2.0
google-cloud-bigquery>=3.0
# Your own product (Streamlit tracker app)
streamlit>=1.28
pandas>=2.0
openpyxl>=3.0
# Optional: Trino for "Refresh from Trino" (same sheet data via Trino/Superset)
trino>=0.328
# Optional: AI section (Ask AI) — set OPENAI_API_KEY in .streamlit/secrets.toml
openai>=1.0
